# Startup time improvements