--- a/src/posting/__init__.py
+++ b/src/posting/__init__.py
@@ -1,3 +1,7 @@
+# This is a hack to prevent httpx from importing _main.py
+import sys
+sys.modules['httpx._main'] = None
+
 from .collection import (
     Auth,
     Cookie,