--- a/src/posting/__init__.py
+++ b/src/posting/__init__.py
@@ -1,3 +1,8 @@
+# This import should be the first thing to run, to ensure that
+# the START_TIME is set as early as possible.
+from posting._start_time import START_TIME  # type: ignore # noqa: F401
+    
+
 # This is a hack to prevent httpx from importing _main.py
 import sys
 sys.modules['httpx._main'] = None