--- a/src/posting/_start_time.py
+++ b/src/posting/_start_time.py
@@ -1,0 +1,4 @@
+import time
+
+
+START_TIME = time.perf_counter()