--- a/src/posting/messages.py
+++ b/src/posting/messages.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
-import httpx
-from textual.events import Message
+from httpx import Response
+from textual.message import Message
 
 
 @dataclass
 class HttpResponseReceived(Message):
-    response: httpx.Response
+    response: Response